chore(project): bump eslint-plugin-bpmn-io to 0.12.0
Bump eslint-plugin-bpmn-io to version 0.12.0 to fix NPM ERESOLVE error.

Version 0.11.0 of lint-plugin-bpmn-io uses eslint versions up to 6.0.0,
but eslint is configured to use 7.14.0. NPM cannot resolve the
dependency and fails to install the modules.

As eslint-plugin-bpmn-io was updated to use eslint up to 7.0.0 with
version 0.12.0 this can be used to fix the issue.