feat: allow to configure custom `nsMap`
This addition now allows for lightweight declaration of additional
important namespaces, to be used to parse a moddle XML document.

One of these is `xmi`, used, i.e. in UML schema definitions. To properly
handle it we must be able to explicitly define the mapping (to it)
without declaring it as an actual Moddle package.

Moddle packages and their `uri -> prefix` mapping still take precedence.